PR review: - fixed lint warnings about rules of hooks - constants separated from hook code - removed 'use' prefix from entities aren't hooks (imported image, style name, event type etc)
