+++ .env.example
@@ -109,5 +109,4 @@ REACT_APP_ADDTOHOME_DESKTOP=
 # Optional (default https://lipis.github.io/flag-icon-css/flags/4x3/XX.svg, XX - country code)
 REACT_APP_FLAGS_URL=
 
-#Count claim for feed
-REACT_APP_ENV_COUNT_CLAIM_FOR_FEED
+REACT_APP_ENV_PHASE_ZERO=
