+++ src/components/common/modal/ModalTopImage.js
@@ -5,7 +5,7 @@ import receiveIllustation from '../../../assets/Feed/receive.svg'
 import sendIllustration from '../../../assets/Feed/send.svg'
 import messageIllustration from '../../../assets/Feed/message.png'
 import inviteIllustration from '../../../assets/Feed/invite.png'
-import useСlaimIllustration from '../../../assets/Feed/IllustrationsMenHero.svg'
+import claimIllustration from '../../../assets/Feed/IllustrationsMenHero.svg' // eslint-disable-line
 import inviteFriendsIllustration from '../../../assets/Feed/inviteFriends.png'
 import backupIllustration from '../../../assets/Feed/backup.png'
 import spendingIllustration from '../../../assets/Feed/spending.svg'
@@ -68,9 +68,9 @@ export const getImageByType = (type, styles = {}) =>
       style: styles.mainPhoto,
       containerStyle: styles.mainPhotoContainer,
     },
-    useclaim: {
-      src: useСlaimIllustration,
-      style: styles.useСlaimIllustrationPhoto,
+    claimsThreshold: {
+      src: claimIllustration,
+      style: styles.claimIllustration,
       containerStyle: styles.mainPhotoContainer,
     },
     backup: {
@@ -134,7 +134,7 @@ const getStylesFromProps = ({ theme }) => ({
     height: '20vh',
     width: '100%',
   },
-  useСlaimIllustrationPhoto: {
+  claimIllustration: {
     marginTop: getDesignRelativeHeight(30),
     height: '24vh',
     width: '60%',
