+++ src/components/dashboard/Claim.js
@@ -26,7 +26,7 @@ import Config from '../../config/config'
 import ClaimAnimatedButton from '../common/animations/ClaimButton/ClaimButton'
 import type { DashboardProps } from './Dashboard'
 import ClaimButton from './ClaimButton'
-import { useClaimCounter } from './Claim/Counter'
+import useClaimCounter from './Claim/useClaimCounter'
 
 type ClaimProps = DashboardProps
 type ClaimState = {
@@ -59,9 +59,9 @@ const Claim = props => {
       amount: '--',
     },
   })
-  const wrappedGoodWallet = wrapper(goodWallet, store)
 
-  const [, advanceClaimsCounter] = useClaimCounter()
+  const wrappedGoodWallet = wrapper(goodWallet, store)
+  const advanceClaimsCounter = useClaimCounter()
 
   // if we returned from facerecoginition then the isValid param would be set
   // this happens only on first claim
@@ -216,8 +216,9 @@ const Claim = props => {
 
       if (receipt.status) {
         fireEvent(CLAIM_SUCCESS, { txhash: receipt.transactionHash })
-        advanceClaimsCounter && (await advanceClaimsCounter())
+        await advanceClaimsCounter()
         checkHanukaBonusDates()
+
         showDialog({
           buttons: [{ text: 'Yay!' }],
           message: `You've claimed your daily G$\nsee you tomorrow.`,
