+++ src/config/config.js
@@ -5,7 +5,7 @@ const Config = {
   env: process.env.REACT_APP_ENV || 'development',
   version: process.env.VERSION || 'v0',
   contractsVersion,
-  isPhaseZero: process.env.REACT_APP_ENV_PHASE_ZERO || true,
+  isPhaseZero: 'true' === process.env.REACT_APP_ENV_PHASE_ZERO,
   newVersionUrl: process.env.REACT_APP_NEW_VERSION_URL || 'https://gdlr.info/newversion',
   logLevel: process.env.REACT_APP_LOG_LEVEL || 'debug',
   serverUrl: process.env.REACT_APP_SERVER_URL || 'http://localhost:3003',
