+++ src/lib/gundb/__tests__/UserStorage.js
@@ -18,7 +18,7 @@ import {
 } from '../UserStorageClass'
 import UserPropertiesClass from '../UserPropertiesClass'
 import { getUserModel } from '../UserModel'
-import { longUseOfClaims } from '../../../components/dashboard/Claim/Counter'
+import { longUseOfClaims } from '../../../components/dashboard/Claim/events'
 import update from '../../updates'
 import { addUser } from './__util__/index'
